requirements.txt
streamlit
mysql-connector-python
bcrypt